Moved tests to KotlinTypeMapping, and polished.
